+++ lib/CalDAV/Backend/PDO.php
@@ -195,8 +195,8 @@ class PDO extends AbstractBackend implements SyncSupport, SubscriptionSupport, S
     /**
      * Creates a new calendar for a principal.
      *
-     * If the creation was a success, an id must be returned that can be used to reference
-     * this calendar in other methods, such as updateCalendar
+     * If the creation was a success, an id must be returned that can be used
+     * to reference this calendar in other methods, such as updateCalendar.
      *
      * @param string $principalUri
      * @param string $calendarUri
@@ -879,7 +879,7 @@ class PDO extends AbstractBackend implements SyncSupport, SubscriptionSupport, S
      */
     protected function addChange($calendarId, $objectUri, $operation) {
 
-        $stmt = $this->pdo->prepare('INSERT INTO ' . $this->calendarChangesTableName .' (uri, synctoken, calendarid, operation) SELECT ?, synctoken, ?, ? FROM calendars WHERE id = ?');
+        $stmt = $this->pdo->prepare('INSERT INTO ' . $this->calendarChangesTableName .' (uri, synctoken, calendarid, operation) SELECT ?, synctoken, ?, ? FROM ' . $this->calendarTableName .' WHERE id = ?');
         $stmt->execute([
             $objectUri,
             $calendarId,